Revert "Use high-resolution timebases for select()."
This reverts commit 6e70fd315a29578f090a98f84e9d8e95ddb156e3.

Revert accidental push

+++ winsup/cygwin/select.h
@@ -96,7 +96,7 @@ public:
 
   bool test_and_set (int, fd_set *, fd_set *, fd_set *);
   int poll (fd_set *, fd_set *, fd_set *);
-  wait_states wait (fd_set *, fd_set *, fd_set *, LONGLONG);
+  wait_states wait (fd_set *, fd_set *, fd_set *, DWORD);
   void cleanup ();
   void destroy ();
